Feature/populate database with omdb data (#22)
* build: database
populate database with data

* setup: database
populating database with data
schema defining using marshamallow
new function for db count
new custom exceptions
custom marshamallow fields
modifed movie model field types
new utils module
request util for querying the OMDB API
config changes for pylint and mypy
git ignore all file types ending with `.db`
fixing the test cases

* setup:  pre-commit-action.yaml
updating mysql config

* test: test cases
test cases for schema
test cases for `utils.build_db`
test cases for `utils.request`

+++ tests/conftest.py
@@ -6,7 +6,8 @@ from flask import Flask, Response
 from flask.testing import FlaskClient
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import Column, Integer
-from sqlalchemy_utils import drop_database
+from sqlalchemy.orm import close_all_sessions
+from sqlalchemy_utils import database_exists, drop_database
 
 from omdb import app
 from omdb.db.base import StringColumn, db
@@ -51,6 +52,7 @@ class BaseTestFlaskClient(FlaskClient):
 
 class BaseTest:
     _app: Flask
+    assert_raises = staticmethod(pytest.raises)
 
     @pytest.fixture
     def test_app(self) -> Flask:
@@ -67,10 +69,11 @@ class BaseTest:
     def init_db(self, test_app: Flask) -> SQLAlchemy:
         with test_app.app_context():
             yield db
-            # TODO: Remove commit after it's handled in after_request
-            db.session.commit()
+            db.session.rollback()
             db.drop_all()
-            drop_database(test_app.config.get('SQLALCHEMY_DATABASE_URI'))
+            close_all_sessions()
+            if database_exists(test_app.config.get('SQLALCHEMY_DATABASE_URI')):
+                drop_database(test_app.config.get('SQLALCHEMY_DATABASE_URI'))
 
     @property
     def client(self) -> BaseTestFlaskClient:
